Incapsulated model from Indicator. This is also fixes issue #1
