+++ .gitignore
@@ -4,3 +4,4 @@
 !*/
 !LuaIndicators/*
 !QLibin/*
+!QLibin/models/*
