+++ LuaIndicators/Price.lua
@@ -1,42 +1,19 @@
-local utils = require("QLibin\\utils")
-local Class = require("QLibin\\oop")
-local Base = require("LuaIndicators\\base")
+local Price = require("QLibin\\models\\Price")
 local QSettings = dofile("QLibin\\settings.lua") -- dofile чтобы
     -- разные индикаторы имели разный экземпляр QSettings
 
 
 Settings = QSettings({
-Name = "*MA = Moving Average2",
+Name = "*Price",
 line = {
     {
-        Name = "MA line",
+        Name = "Price line",
         Type = TYPE_LINE,
         Color = RGB(180, 170, 160)
         },
     },
-period = 5
+hloc = 'C'
 })
-
--- utils.log(Settings)
-
-local function getHLOCFunc(hloc)
-    if hloc == 'H' then return H end
-    if hloc == 'L' then return L end
-    if hloc == 'O' then return O end
-    if hloc == 'C' then return C end
-    utils.log(utils.dump(hloc)..' not found in {H,L,O,C}. Using C instead.')
-    return C -- default will be Close
-end
-
-local Price = Class(Base)
-
-function Price:calc(index)
-	if index > 0 then
-        return getHLOCFunc(self.settings.hloc)(index)
-    else
-        return nil
-    end
-end
 
 local indicator
 
@@ -48,5 +25,3 @@ end
 function OnCalculate(Index)
     return indicator(Index)
 end
-
-return Price
