+++ QLibin/models/fractals_strategy.lua
